Revert to static scheduling
There's a race condition in overlord where xmls dropped into the kiosk
folder can be read before the file lock is released. Fixing this
necessitated introducing a separate script to perform atomic moves into
the kiosk folder to avoid triggering the race condition. Somehow, this
triggered a new bug where overlord would repeatedly queue the same xml
file. It seems there's no winning.

The only way to avoid these bugs is to constrain the time in which
synapse interacts with the kiosk folder to the few minutes after
overlord starts up. If the bug is triggered, the damage is minimal.